+++ drivers/staging/android/ion/mtk/ion_mm_heap.c
@@ -1186,7 +1186,7 @@ struct ion_heap *ion_mm_heap_create(struct ion_platform_heap *unused)
 		return ERR_PTR(-ENOMEM);
 	}
 	heap->heap.ops = &ion_mm_heap_ops;
-	heap->heap.type = ION_HEAP_TYPE_MULTIMEDIA;
+	heap->heap.type = (unsigned int)ION_HEAP_TYPE_MULTIMEDIA;
 	heap->heap.flags = ION_HEAP_FLAG_DEFER_FREE;
 	heap->pools =
 	    kcalloc(num_orders, sizeof(struct ion_page_pool *), GFP_KERNEL);
